\name{lab5webAPI-package}
\alias{lab5webAPI-package}
\alias{lab5webAPI}
\docType{package}
\title{
\packageTitle{lab5webAPI}
}
\description{
\packageDescription{lab5webAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5webAPI}
\packageIndices{lab5webAPI}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{lab5webAPI}

Maintainer: \packageMaintainer{lab5webAPI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
